refactor tasks table
